update issue #23, #4 Added draft implementation of import progress functionality into import wizard. Some other refactoring

+++ trunk/Modules.Import/Converters/ProgressPercentageConverter.cs
@@ -10,7 +10,7 @@ namespace Modules.Import.Converters
 
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            double dval = (double)value;
+            float dval = (float)value;
             return String.Format("{0:0.00}", dval) + "%";
         }
